+++ library/ClickHaskell.hs
@@ -63,6 +63,7 @@ module ClickHaskell
   {- * ClickHouse types -}
   , IsChType(chTypeName, defaultValueOfTypeName)
   , DateTime, DateTime64
+  , Date
   , Int8, Int16, Int32, Int64, Int128(..), Int256(..)
   , UInt8, UInt16, UInt32, UInt64, UInt128, UInt256, Word128(..), Word256(..)
   , Float32, Float64